Refactoring and adding generic exercise form
