+++ lib/src/features/core/screens/profile/profile_screen.dart
@@ -1,5 +1,6 @@
 import 'package:fit_office/src/features/core/controllers/exercise_timer.dart';
 import 'package:fit_office/src/features/core/screens/dashboard/widgets/active_dialog.dart';
+import 'package:fit_office/src/features/core/screens/profile/admin/exercise_form.dart';
 import 'package:fit_office/src/features/core/screens/profile/widgets/facet_display_card.dart';
 import 'package:fit_office/src/utils/helper/dialog_helper.dart';
 import 'package:flutter/material.dart';
@@ -8,11 +9,9 @@ import 'package:line_awesome_flutter/line_awesome_flutter.dart';
 import 'package:fit_office/src/constants/sizes.dart';
 import 'package:fit_office/src/constants/text_strings.dart';
 import 'package:fit_office/src/features/core/screens/profile/widgets/update_profile_modal.dart';
-import '../../../../common_widgets/buttons/primary_button.dart';
 import '../../../../constants/colors.dart';
 import '../../../../repository/authentication_repository/authentication_repository.dart';
 import '../../controllers/profile_controller.dart';
-import 'admin/add_exercises.dart';
 import 'admin/add_friends.dart';
 import 'admin/edit_user_page.dart';
 import 'admin/widgets/all_users.dart';
@@ -338,7 +337,7 @@ class ProfileScreen extends StatelessWidget {
                           return;
                         }
 
-                        Get.to(() => AddExercises(currentUserId: user.id!));
+                        Get.to(() => ExerciseForm(isEdit: false));
                       },
                     ),
                     CustomProfileButton(
